% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{add_to_selected_mitigators}
\alias{add_to_selected_mitigators}
\title{Add to the selected mitigator list}
\usage{
add_to_selected_mitigators(df, selected_currently, new_selections)
}
\arguments{
\item{df}{Tibble of mitigator reference data - e.g. `mitigator_reference`}

\item{selected_currently}{A list of mitigator_codes which are currently
selected - as provided by `input$mitigators`}

\item{new_selections}{A list of newly selected mitigatoe}
}
\value{
Vector - mitigator labels with mitigator codes as the value
}
\description{
Adds newly selected mitigators to the list of currently selected mitigators
}
