% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_distribution_characteristics}
\alias{get_distribution_characteristics}
\title{Summarise mixture distributions.}
\usage{
get_distribution_characteristics(normal_dists, mix_dists, mitigators)
}
\arguments{
\item{normal_dists}{A dataframe with the mu and sigma of the normal
distribution for each peer and mitigator.}

\item{mix_dists}{A list of mixture distributions for each mitigator.}

\item{mitigators}{A vector of the unique mitigators.}
}
\value{
A dataframe of distribution characteristics.
}
\description{
Get mu, sigma, number of peers, p10, p50 and p90 for each mixture
distribution.
}
