% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_mixture_distributions}
\alias{get_mixture_distributions}
\title{Create mixture distributions for each mitigator:}
\usage{
get_mixture_distributions(data, mitigators)
}
\arguments{
\item{data}{A dataframe of with the mu and sigma of the normal distribution
for each peer and mitigator.}

\item{mitigators}{A vector of the unique mitigators.}
}
\value{
A list of mixture distributions for each mitigator.
}
\description{
Aggregates over the peers to create a mixture distribution for each activity
subset.
}
