% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_mixture_distributions_dat}
\alias{get_mixture_distributions_dat}
\title{Get mixture distribution data}
\usage{
get_mixture_distributions_dat(dat)
}
\arguments{
\item{dat}{Tibble - of mitigator data as produced by `populate_table()` in `fct_tabulate.R`}
}
\value{
Tibble of mixture distribution data for each mitigator
}
\description{
Co-ordinates the production of mixture distribution data from the various
functions above.
}
