% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_mu_sigma}
\alias{get_mu_sigma}
\title{Get mu and sigma from aggregating over normal distributions.}
\usage{
get_mu_sigma(data)
}
\arguments{
\item{data}{A dataframe with the mu and sigma of the normal distribution
for each peer and mitigator.}
}
\value{
A dataframe with mu, sigma and number of peers for each mixture
distribution.

Note mean and sd of unweighted mixture distribution of Normal distributions
is:
mu_mix = sum(mu) / n
sd_mix = ( sum(mu^2 + sigma^2) / n ) ^ (1/2)
source : https://stats.stackexchange.com/questions/447626/mean-and-variance-of-a-mixture-distribution
}
\description{
Get mu and sigma from aggregating over normal distributions.
}
