% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_p10_p50_p90}
\alias{get_p10_p50_p90}
\title{Get p10, p50 and p90 of mixture distributions.}
\usage{
get_p10_p50_p90(data, mitigators)
}
\arguments{
\item{data}{A list of mixture distributions for each mitigator.}

\item{mitigators}{A vector of the unique mitigators.}
}
\value{
A dataframe with p10, p50 and p90 of each mixture distribution.
}
\description{
Get p10, p50 and p90 of mixture distributions.
}
