% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_percentiles}
\alias{get_percentiles}
\title{Get percentiles from the mixture distributions.}
\usage{
get_percentiles(data, mitigators)
}
\arguments{
\item{data}{A list of mixture distributions for each mitigator.}

\item{mitigators}{A vector of the unique mitigators.}
}
\value{
A long dataframe of the percentiles for the ECDF and PDF of each
mitigator's mixture distribution.
}
\description{
Gets the percentiles for the ECDF and PDF of each mitigator's mixture
distribution.
}
