% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{get_probability_plot}
\alias{get_probability_plot}
\title{Probability plot.}
\usage{
get_probability_plot(data, type)
}
\arguments{
\item{data}{A dataframe where each row is a percentile of the ECDF and PDF of
a mixture distribution. Distribution characteristics and labels, groups are
also included.}

\item{type}{Either `"ecdf"` or `"pdf"` to get the empirical cumulative
distribution functions or probability density functions, respectively.}
}
\value{
A plot of the ECDF or PDF.
}
\description{
Probability plot.
}
