% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{get_trust_lookup}
\alias{get_trust_lookup}
\title{Get a lookup table of participating Trusts}
\usage{
get_trust_lookup(container_support)
}
\arguments{
\item{container_support}{The Azure container for supporting information, as
obtained by `get_container()` from `fct_azure.R`.}
}
\value{
Tibble of data listing participating Trusts
}
\description{
Read a csv lookup file from Azure storage and do some clean-up to ensure
one row per Trust.
}
