% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lookups.R
\name{make_mitigator_uptake_dat}
\alias{make_mitigator_uptake_dat}
\title{Prepare data for the mitigator uptake DT object}
\usage{
make_mitigator_uptake_dat(dat, selected_schemes)
}
\arguments{
\item{dat}{Tibble - the full prepared dataset for this app}

\item{selected_schemes}{Character vector - a list of mitigator_codes selected by the user}
}
\value{
Tibble listing the mitigators and summaries of schemes using them
}
\description{
Renders a tibble showing the proportion of schemes using each mitigator.
Two rates are shown:
1. covers all available schemes,
2. covers the subset of schemes selected by the user.
}
