% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lookups.R
\name{make_scheme_dt}
\alias{make_scheme_dt}
\title{Make the scheme lookup DT object}
\usage{
make_scheme_dt(trust_code_lookup)
}
\arguments{
\item{trust_code_lookup}{Tibble of scheme data - as provided by the
`get_trust_lookup()` function in `fct_tabulate.R`}
}
\value{
DT object listing schemes
}
\description{
Renders a DT object listing schemes represented in the app with filters and a
CSV download button.
}
