% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lookups.R
\name{make_scheme_uptake_dat}
\alias{make_scheme_uptake_dat}
\title{Prepare the data for the scheme mitigator uptake DT object}
\usage{
make_scheme_uptake_dat(
  dat,
  selected_mitigators,
  selected_schemes,
  focal_scheme
)
}
\arguments{
\item{dat}{Tibble - the full prepared dataset for this app}

\item{selected_schemes}{Character vector - a list of scheme_codes selected by the user}

\item{focal_scheme}{Character vector - the focal scheme_code}
}
\value{
Tibble listing schemes and the proportions of mitigators in use by them
}
\description{
Renders a tibble showing the proportion of mitigators used by each scheme.
Two rates are shown:
1. covers all available mitigators,
2. covers the subset of mitigators selected by the user.
}
