% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{modify_theme}
\alias{modify_theme}
\title{Modifies theme of ECDF and PDF plot.}
\usage{
modify_theme(plot, type)
}
\arguments{
\item{plot}{A plot of an ECDF or PDF.}

\item{type}{Either `"ecdf"` or `"pdf"` to get the modifications for the
empirical cumulative distribution functions or probability density
functions, respectively.}
}
\value{
A plot with theme modifiers based on whether the plot was is an ECDF
  or PDF.
}
\description{
Modifies theme of ECDF and PDF plot.
}
