% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{percent_breaks}
\alias{percent_breaks}
\title{Set breaks for a percent value}
\usage{
percent_breaks(n = 5, ...)
}
\arguments{
\item{n}{integer - the approximate number of breaks to set (defaults to 5)}

\item{...}{other parameters, not yet defined}
}
\value{
ggplot2::breaks object
}
\description{
An internal function to produce ggplot breaks on 0.01 (1%) intervals.
Based on code found here:
https://jhrcook.github.io/jhrcook-website/posts/2019-11-09_integer-values-ggplot-axis/
}
