% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_faceted_trendlines}
\alias{plot_faceted_trendlines}
\title{Plot faceted trendlines}
\usage{
plot_faceted_trendlines(
  dat,
  rates_data,
  mitigator_codes,
  focal_scheme_code,
  scheme_codes,
  show_other_schemes = TRUE,
  show_horizon_timeline = TRUE,
  show_horizon_overlay = TRUE,
  show_prebaseline_average = TRUE,
  facet_height_px = 400,
  return_data = FALSE
)
}
\arguments{
\item{dat}{Tibble - mitigator data as produced by `populate_table()` in
`fct_tabulate.R`}

\item{rates_data}{Tibble - historical rates data for schemes / mitigators}

\item{mitigator_codes}{Character vector - the mitigator codes to visualise}

\item{focal_scheme_code}{Character vector - the scheme code for the focal
scheme}

\item{scheme_codes}{Character vector - a list of codes for other schemes to
visualise}

\item{show_other_schemes}{Boolean (default = TRUE) plot time series for other
schemes}

\item{show_horizon_timeline}{Boolean (default = TRUE) plot the predicted
activity on the timeline at the horizon year}

\item{show_horizon_overlay}{Boolean (default = TRUE) plot the predicted
activity as an overlay over the historical time series plots}

\item{show_prebaseline_average}{Boolean (default = TRUE) show the
pre-baseline average (mean) with a range of two standard deviations above
and below}

\item{facet_height_px}{Integer (default = 400) the height of the plot}

\item{return_data}{Boolean (default = FALSE) TRUE = return a list object of
tibbles used in the production of the plot - FOR TROUBLESHOOTING PURPOSES
ONLY}
}
\value{
{plotly} plot combining individual trendline plots for each
  mitigator
}
\description{
Co-ordinates the production of a a \{plotly\} object containing multiple time
series plots, one for each mitigator in `mitigator_codes`.
}
