% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap plot}
\usage{
plot_heatmap(dat_selected_heatmap, input)
}
\arguments{
\item{input}{Reference to the Shiny input widget that triggered this chart}

\item{dat_selected_pointrange}{Tibble of mitigator data as produced by `populate_table()` in `fct_tabulate.R`}
}
\value{
ggplot2 object showing heatmap
}
\description{
Construct the heatmap showing values for mitigators when compared across
schemes.
}
\details{
Options include:

1. Mitigator name
   Checkbox-controlled switch between mitigator code and mitigator name to
   help interpret the plots.

2. Value to plot
   Drop-down selection of binary (if a mitigator is used at all), value, low,
   high and range

x-axis responsivity:
* <6 schemes selected - names are displayed vertically and wrapped at 20 char
* 6 <= x <12 schemes - names are displayed at 45 degrees with no wrap
* >= 12 schemes - names are rotated at 90 degrees

y-axis:
* string-wrapped at around half the length of the longest mitigator name
}
