% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_pointrange}
\alias{plot_pointrange}
\title{Pointrange plots}
\usage{
plot_pointrange(dat_selected_pointrange, input)
}
\arguments{
\item{dat_selected_pointrange}{Tibble of mitigator data as produced by
`populate_table()` in `fct_tabulate.R`}

\item{input}{Reference to the Shiny input widget that triggered this chart}
}
\value{
ggplot2 object showing point-range view facetted by scheme or
  mitigator
}
\description{
Construct the point-range
}
\details{
Options include:
1. Standardised horizon
   Divides the mitigator by years from start to final year.

2. Mitigator code
   Replaces the mitigator name with the mitigator code - to reduce visual
   cutter.

3. Facet by scheme
   By default the mitigators are faceted with schemes on the y-axis. This
   option plots the mitigators on the y-axis instead and the plot is faceted
   by scheme.

4. Facet rows
   Allows the user to define how many rows to facet the plot over.

5. Show NEE reference range
   Allows the user to display the range of values from the National Elicitation
   Exercise (NEE) as context. Defaults to on.
}
