% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{prepare_mitigators}
\alias{prepare_mitigators}
\title{Wrangle Mitigator Lookup}
\usage{
prepare_mitigators(mitigator_lookup)
}
\arguments{
\item{mitigator_lookup}{Tibble of mitigator lookup data}
}
\value{
Tibble of mitigator lookups
}
\description{
Prepare mitigator lookup file that's been read in from Azure. Rename/adjust
content to recreate the original form of the mitigator lookup file, which has
since been adjusted.
}
