% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure.R
\name{read_provider_data}
\alias{read_provider_data}
\title{Read Rates (Trend) Data
Fetch and read parquet files that store contextual data used in the NHP
inputs app.}
\usage{
read_provider_data(
  container_inputs,
  file = "rates",
  inputs_data_version = Sys.getenv("NHP_INPUTS_DATA_VERSION", "dev")
)
}
\arguments{
\item{container_inputs}{A storage container object.}

\item{file}{Character. Name of parquet file to be read (without the
extension).}

\item{inputs_data_version}{Character. The app version for which you want to
return data (there's a separate versioned folder of parquet files for
each release).}
}
\value{
A tibble.
}
\description{
Read Rates (Trend) Data
Fetch and read parquet files that store contextual data used in the NHP
inputs app.
}
