% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamods_select_group.R
\name{select-group}
\alias{select-group}
\alias{select_group_ui}
\alias{select_group_server}
\title{Select Group Input Module}
\usage{
select_group_ui(
  id,
  params,
  label = NULL,
  btn_reset_label = "Reset filters",
  inline = TRUE,
  vs_args = list()
)

select_group_server(id, data_r, vars_r)
}
\arguments{
\item{id}{Module's id.}

\item{params}{A list of parameters passed to each [shinyWidgets::virtualSelectInput()],
you can use :
 * `inputId`: mandatory, must correspond to variable name.
 * `label`: Display label for the control.
 * `placeholder`: Text to show when no options selected.}

\item{label}{Character, global label on top of all labels.}

\item{btn_reset_label}{Character, reset button label. If `NULL` no button is added.}

\item{inline}{If `TRUE` (the default),
select menus are horizontally positioned, otherwise vertically.}

\item{vs_args}{Arguments passed to all [shinyWidgets::virtualSelectInput()] created.}

\item{data_r}{Either a [data.frame()] or a [shiny::reactive()]
function returning a `data.frame` (do not use parentheses).}

\item{vars_r}{character, columns to use to create filters,
must correspond to variables listed in `params`. Can be a
[shiny::reactive()] function, but values must be included in the initial ones (in `params`).}
}
\value{
A [shiny::reactive()] function containing data filtered with an attribute `inputs` containing a named list of selected inputs.
}
\description{
Group of mutually dependent select menus for filtering `data.frame`'s columns (like in Excel).
}
