% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{update_dat_values}
\alias{update_dat_values}
\title{Update dat to reflect the user's preferred values}
\usage{
update_dat_values(dat, values_displayed)
}
\arguments{
\item{dat}{Tibble of data - as produced from `populate_table` in `fct_tabulate.R`}

\item{values_displayed}{Character - the value in `input$values_displayed` indicating the user's preferred view}
}
\value{
Tibble of data with the 'value_' and 'nee_' fields updated to match the user's preferred values
}
\description{
Users are able to select from 'Percent of activity mitigated' and '80%
prediction interval' in the global settings sidebar.
}
\details{
This function updates dat to reflect the user's preferred values.
}
