% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mixture_distributions.R
\name{wrangle_data_for_density_plots}
\alias{wrangle_data_for_density_plots}
\title{Prepare data for plotting densities.}
\usage{
wrangle_data_for_density_plots(
  peer_agg_ecdf_pdf,
  peer_agg_dist_summary,
  strategy_lookup
)
}
\arguments{
\item{peer_agg_ecdf_pdf}{A long dataframe of the percentiles for the ECDF and
PDF of each mitigator's mixture distribution.}

\item{peer_agg_dist_summary}{A dataframe of distribution characteristics.}

\item{strategy_lookup}{A dataframe of labels, types and groups for each
strategy.}
}
\value{
Tibble of data showing mixture distributions
}
\description{
Wrangles data to a dataframe for plotting ECDFs and PDFs
}
