% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{wrap_strings_to_fit_pixel_limit}
\alias{wrap_strings_to_fit_pixel_limit}
\title{Wrap a vector of strings to fit a specified width}
\usage{
wrap_strings_to_fit_pixel_limit(
  strings,
  px_limit = 300,
  font_family = "Arial, Helvetica, Droid Sans, sans",
  font_size = 14
)
}
\arguments{
\item{strings}{Character vector - the input vector of strings to wrap}

\item{px_limit}{Integer (default 300) - the number of pixels to wrap `strings` to fit}

\item{font_family}{Character (default 'Arial, Helvetica, Droid Sans, sans') - the font family to use when estimating the width of strings}

\item{font_size}{Integer (default 14) - the font size to use when estimating the width of strings}
}
\value{
Character vector
}
\description{
Takes a vector of strings and wraps them to fit within a specified pixel
width limit.
}
