% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{populate_table}
\alias{populate_table}
\title{Populate dat table}
\usage{
populate_table(
  skeleton_table,
  extracted_params,
  trust_code_lookup,
  mitigator_lookup,
  nee_results
)
}
\arguments{
\item{skeleton_table}{Tibble - output of `prepare_skeleton_table()`}

\item{extracted_params}{Tibble - output of `extract_params()`}

\item{trust_code_lookup}{Tibble - from Azure file `nhp-scheme-lookup.csv`}

\item{mitigator_lookup}{Tibble - from Azure file `mitigator-lookup.csv`}

\item{nee_results}{Tibble - from fct_read.R of `nee_table.rds`}
}
\value{
Tibble of data
}
\description{
Creates a single fact table from `skeleton_table` by left-joining with
`extracted_params`, `trust_code_lookup`, `mitigator_lookup` and
`nee_results`.
}
