% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{prepare_mitigators}
\alias{prepare_mitigators}
\title{Prepare Mitigator lookup}
\usage{
prepare_mitigators(mitigator_lookup)
}
\arguments{
\item{mitigator_lookup}{Tibble of mitigator lookup data}
}
\value{
Tibble of mitigator lookups
}
\description{
Prepare the mitigator lookup file for subsequent use by
a) cleaning variable names
b) ordering the variables by cardinality
}
