% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tabulate.R
\name{update_dat_values}
\alias{update_dat_values}
\title{Update dat to reflect the user's preferred values}
\usage{
update_dat_values(
  dat,
  values_displayed,
  include_point_estimates = FALSE,
  focal_scheme_code
)
}
\arguments{
\item{dat}{Tibble of data - as produced from `populate_table` in `fct_tabulate.R`}

\item{values_displayed}{Character - the value in `input$values_displayed` indicating the user's preferred view}

\item{include_point_estimates}{Boolean - the value in `input$include_point_estimates` indicating whether point-estimates of 0% mitigation (100% prediction) are to be included}

\item{focal_scheme_code}{Character - the scheme code of the focal scheme}
}
\value{
Tibble of data with the 'value_' and 'nee_' fields updated to match the user's preferred values
}
\description{
Users are able to select from 'Percent of activity mitigated' and '80%
prediction interval' in the global settings sidebar.
}
\details{
This function updates dat to reflect the user's preferred values.
}
